BindGroupSource trait; cleaned up code
I added a BindGroupSource trait because I feel like it'll make it neater to add data to send to shaders.

Most of it was me moving the Mesh types to lib.rs, generalizing them to Mesh<V> (V is a vertex type), and shifting some code to util functions in lib.rs.

Not sure if this was all necessary but it might make things easier.